requirements.txt
flask
pandas
numpy
scikit-learn
joblib
psycopg2-binary
python-dotenv
xgboost
